perf: optimize directory structure
