+++ .prettierrc
@@ -1,9 +1,9 @@
 {
-	"useTabs": true,
-	"printWidth": 80,
-	"tabWidth": 2,
-	"singleQuote": true,
-	"trailingComma": "none",
+  "useTabs": false,
+  "tabWidth": 2,
+  "printWidth": 88,
+  "singleQuote": true,
+  "trailingComma": "none",
   "bracketSpacing": true,
   "semi": false
 }
