+++ src/components/Main/index.styl
@@ -1,0 +1,67 @@
+@import "../../style.styl"
+
+$top-height = 60px
+$left-side-width = 80px
+
+.main-container {
+  position absolute
+  box-sizing border-box
+  width 100%
+  height 100%
+  background $second-background-color
+  padding-top $top-height
+
+
+  .top {
+    position absolute
+    top 0
+    left 0
+    box-sizing border-box
+    width 100%
+    height $top-height
+  }
+
+
+  .bottom {
+    //position relative
+    //box-sizing border-box
+    //width 100%
+    //height 100%
+    //display flex
+    //justify-content space-between
+    //padding-top $component-interval
+
+    position relative
+    box-sizing border-box
+    width 100%
+    height 100%
+    padding-top $component-interval
+    padding-left $left-side-width
+
+
+    .left {
+      position absolute
+      width $left-side-width
+      height 100%
+      top 0
+      left 0
+      box-sizing border-box
+    }
+
+
+    .right {
+      position relative
+      width 100%
+      height 100%
+      box-sizing border-box
+      padding 0 $component-interval 0 $component-interval
+
+      .content {
+        position relative
+        box-sizing border-box
+        width 100%
+        height 100%
+      }
+    }
+  }
+}
