--- src/components/Step1.vue
@@ -1,17 +1,0 @@
-<template>
-  <div>
-    <h3>创建一个用来存储图片的<a href="https://github.com/new" target="_black" onclick="event.stopPropagation()">GitHub仓库</a></h3>
-    <img
-      src="https://cdn.jsdelivr.net/gh/Coffee-kitten/tu/images/repo1.png"
-      alt="img"
-      style="zoom: 50%"
-    />
-  </div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style>
-</style>
