--- src/components/Step2.vue
@@ -1,26 +1,0 @@
-<template>
-  <div>
-    <h3>创建一个有 repo 权限的<a href="https://github.com/settings/tokens/new" target="_black" onclick="event.stopPropagation()">GitHub Token</a></h3>
-    <img
-      src="https://cdn.jsdelivr.net/gh/Coffee-kitten/tu/images/token.png"
-      alt="img"
-      style="zoom: 50%"
-    />
-    <p>然后点击 Generate token 按钮，即可生成一个token，如下：</p>
-    <img
-      src="https://cdn.jsdelivr.net/gh/WishMelz/file/image/token.png"
-      alt=""
-    />
-    <p style="color: red"><em>​ 生成的TOKEN只会显示一次。记得保存！！</em></p>
-  </div>
-</template>
-
-<script>
-export default {
-
-}
-</script>
-
-<style>
-
-</style>
