--- src/components/Step3.vue
@@ -1,46 +1,0 @@
-<template>
-  <div>
-    <h3>进行图床配置（绑定 GitHub Token、存储图片的仓库和目录）</h3>
-    <p>先设置用户token和仓库路径</p>
-
-    <img
-      src="https://cdn.jsdelivr.net/gh/Coffee-kitten/tu/images/use.jfif"
-      alt="img"
-      style="zoom: 60%"
-    />
-
-    <p></p>
-
-    <img
-      src="https://cdn.jsdelivr.net/gh/Coffee-kitten/tu/images/use1.png"
-      alt="img"
-      style="zoom: 30%"
-    />
-
-    <p>最后一步</p>
-
-    <img
-      src="https://cdn.jsdelivr.net/gh/Coffee-kitten/tu/images/use2.jfif"
-      alt="img"
-      style="zoom: 60%"
-    />
-
-    <p style="text-align:left">- 新建目录：需手动输入一个新目录。</p>
-
-    <p style="text-align:left">- 根目录：图片将直接存储在仓库根目录下。</p>
-
-    <p style="text-align:left">- 自动目录：自动生成日期格式 `YYYYMMDD` 的目录。例如：`20200909`。</p>
-
-    <p style="text-align:left">- 选择仓库目录：自动获取仓库下所有目录，选择一个即可。</p>
-
-    <p style="text-align:left">至此完成图床配置，现在就可以开始上传图片了，快去试试吧~</p>
-
-  </div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style>
-</style>
