+++ src/components/site-count.vue
@@ -1,6 +1,7 @@
 <template>
   <span class="site-count" ref="siteCountDom">
-    <span id="busuanzi_value_site_pv"></span>
+    <span id="busuanzi_value_site_uv" v-if="isuv"></span>
+    <span id="busuanzi_value_site_pv" v-if="!isuv"></span>
   </span>
 </template>
 
@@ -8,10 +9,17 @@
 import { defineComponent, onMounted, ref, Ref } from 'vue'
 
 export default defineComponent({
-  name: "site-count",
+  name: 'site-count',
+
+  props: {
+    isuv: {
+      type: Boolean,
+      default: false
+    }
+  },
 
   setup() {
-    const siteCountDom: Ref = ref<null | HTMLElement>(null);
+    const siteCountDom: Ref = ref<null | HTMLElement>(null)
 
     onMounted(() => {
       const script: any = document.createElement('script')
