+++ src/views/About/index.styl
@@ -1,0 +1,25 @@
+.feedback-page-container {
+
+  .help-info-item {
+    font-size: 15.8px;
+    padding: 6px;
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+
+
+  }
+
+  .description {
+    font-weight: bold;
+    line-height: 28px;
+  }
+
+  .red-text {
+    color: #de1a1a;
+  }
+}
