+++ src/views/Config/index.styl
@@ -1,0 +1,13 @@
+.config-page-container {
+  .operation {
+    text-align right
+
+    .el-button {
+      margin-left 20px
+
+      &:first-child {
+        margin-left 0
+      }
+    }
+  }
+}
