+++ src/views/Management/index.styl
@@ -1,0 +1,56 @@
+@import "../../style.styl"
+
+$infoBarHeight = 50px;
+
+.management-page-container {
+
+  .content-container {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    padding-top: $infoBarHeight;
+    box-sizing: border-box;
+
+    .top {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: $infoBarHeight;
+      box-sizing: border-box;
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      font-size: 14px;
+      padding-bottom: 20px;
+    }
+
+    .bottom {
+      width: 100%;
+      height: 100%;
+      box-sizing: border-box;
+
+      .image-list {
+        width: 100%;
+        height: 100%;
+        margin: 0;
+        padding: 0;
+        list-style: none;
+        border: 1px solid $image-list-border-color;
+        overflow-y: auto;
+        box-sizing: border-box;
+
+        li.image-item {
+          float: left;
+          box-sizing: border-box;
+          padding: 10px;
+
+          &:last-child {
+            margin-right: 0;
+          }
+        }
+      }
+
+    }
+  }
+}
