+++ src/views/Tutorials/index.styl
@@ -1,0 +1,15 @@
+.tutorials-page-container {
+
+  .step-content {
+    padding-top 30px
+    display flex
+    justify-content center
+    text-align center
+  }
+
+  .btn-next-prev {
+    text-align center
+    padding-top 30px
+  }
+
+}
