+++ vite.config.ts
@@ -7,19 +7,13 @@ export default defineConfig({
   plugins: [vue()],
   resolve: {
     alias: {
-      '@': resolve(__dirname, 'src'),  // 设置绝对路径别名
+      '@': resolve(__dirname, 'src') // 设置 @ 指向 src
     }
   },
-  base: './',  // 设置打包路径
+  base: './', // 设置打包路径
   server: {
     port: 4000,
     open: true,
-    // proxy: {
-    //   '/api': 'http://xxx.xxx.xxx.xxx:8000' // 设置代理
-    // },
     cors: true
   }
 })
-
-
-
